removed authentication against database in AMOAuthentication fetch_amo_user moved to utils.amo fixed save profile for new users or users without profile
